Changing names so beam works by convention
I didn't feel like altering the beam database settings, so this just makes the
actual names in the db match up with what beam is expecting them to be.

In the tutorial, the table is changes to just "addresses", since beam will by
convention look for "cart_user_addresses"

The same thing happens with "line1" and "line2" in the AddressT table.

This just renames the record fields so that we don't have to do 'dbModification'
in beam and in the code.