continuing with the tutorial
adding lens to make things a little nicer to read

need allAddresses much like we needed allUsers from before

selecting all user and addresses is an inner join, which only works with an on
clause in postgres, so only using related_ works.

Not a fan of mixing lens syntax with normal record syntax, but getting compile
errors trying to make the record syntax be a lens syntax

All the rest of the tutorial examples for joining don't produce an on clause in
the sql so postgres barfs at that syntax, probably need to ask what is going on
with this

Using a guard_ statement to produce a where clause to get all address belonging
to betty.  Finally can get a where clause working, couldn't get it working in
tutorial 1 for some reason