support for order and shipping table
order table has a foreign key to the user who put in the order
and a foreign key to the address of the user who ordered it
the foreign (beam docs tad this as a primary key) key to the shipping
information of the user is optional -- hence the
Nullable wrapper -- any x is a Maybe x with that wrapper.

you probably should install a primary key on the shipping order table and then
you wouldn't need Nullable, but in the tutorial this is used to show some of
beams features.

Likely not useful in anything you create new, but this is present in some
existing databases

need to set up the database settings