using db modification in beam
this lets us be more haskelly with the records, at least with their names and
lenses if we want them, but keeping the sql reasonable and conventional as well.

This is like the marriage of both of these naming conventions.

The data type address is really addresses in the database

the field _addressLine1 would be defaulted in beam to column with the name
line1, but we want our sql column to be named address1 instead.

the same thing is happening to products, orders, shipping info and line items.