inserting users addresses and products with returningList
nifty little destructuring and binding inside a vector at the top level
declarations.

note that we need another import

import Database.Beam.Backend.SQL.BeamExtensions

Since there's no standard insert returning syntax across databases, it's not
imported by default.

Also the syntax for sqlite is different than the syntax for postgres.

As per the tutorial, sqlite syntax is

  withDatabaseDebug putStrLn conn $
    runInsertReturningList $ insertReturning (shoppingCartDb ^. shoppingCartUserAddresses) $
    insertValues addresses

but for postgres the syntax is

  withDatabaseDebug putStrLn conn $
    runInsertReturningList (shoppingCartDb ^. shoppingCartUserAddresses) $
    insertValues addresses

The confusion comes because beam-postgres has its own INSERT ... RETURNING
... syntax exposed by the insertReturning function. This is specific to
postgres.

Also an extra selection for a UserId and not just a Text.  Either way you need
val_ but if you are comparing a full primary key, you need to use the
_addressForUser accessor on the record to get that.  The lens unwraps the data
constructor that accessor points to, leaving you with a Text value basically.
So you can only use the lens syntax when comparing against a Text value, and the
record accessor when comparing against the whole primary key.

You can change this an use template haskell with lens, and that should work, but
I haven't done that yet.