getting the marshalling to work for ShippingCarrier
the tutorial was for sql lite and postgres indeed has an extra parameter it
passes to fromField.

This was present in the error messages with 'too few arguments passed in to
fromField'

was getting a type error 'Maybe bytestring -> Conversion` was found but expected
a Conversion.  That was the hint really